Update MAURConfig.m
This is a bug fis. It's pretty critical and can cause a lot of headache, did for me. You can get road snapping when you don't want to for example, ie Fitness activity type becomes AutomotiveNavigation.

String comparisons in MAURConfig.m look like this now:
if ([activityType caseInsensitiveCompare:@"AutomotiveNavigation"]) {
return CLActivityTypeAutomotiveNavigation;
}

but that actually means "if string NOT matches". It should instead be

if ([activityType caseInsensitiveCompare:@"AutomotiveNavigation"] == NSOrderedSame) {
return CLActivityTypeAutomotiveNavigation;
}